requirements.txt
anyio==3.6.2
asgiref==3.6.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
Django==4.1.7
djangorestframework==3.14.0
requests==2.28.2
gunicorn==20.1.0
psycopg2-binary==2.9.1